Added timeouts for Variation and Deviation
Previously, if a variation or deviation value was ever received, it was used for the entirety of the program's life. Now, those data expire after 4 sec, which should be within the window of that data being received on n2k.
This was noticed when a seemingly errant value for Deviation was received, and set to 186deg. The true heading was then wrong until program reset.